Extract frontend library and playback hooks
